Refactor, the first of many
Since the yace project is getting more complicated I opted to go with a build system,
namely CMake. I didn't have much experience with CMake so now might be a good time to start.
Need to learn how to automate run files though, but for now this is a good starting point.

I also added framework for tests, which should be coming soon.